Widgets: remove unnecessary code from list box demo
The logic unchecks all other rows whenever a row is checked.
The same could be achieved by setting multiCheck to false. So, the logic
is implemented too complicated and could be simplified.
However, the default behavior of a list box allows multi check, so it is
confusing if this specific example box does not -> it is better to
remove the code all together.

385612