Do not package js files to jar for library modules
A library module (i.e. a module that uses libraryConfig() in its
webpack.config.js) should not include the webpack output in the JAR
file. Instead, the webpack output is published as NPM artifact. Only
bundle modules (usually called *.app) include the final webpack output
in their JAR file, so the sources are available on the class path, but
only once (no duplicates).