UiNotifications: wait for subscribe
The subscribe request needs to be awaited to avoid race conditions and
to improve error handling.

Example what happens without waiting for subscribe():
1. Call subscribe(). A notification for that topic would reload a local
   cache.
2. Subscribe request is still pending.
3. Load data to fill a local cache.
4. Publish a notification to invalidate the cache.
4. Subscribe request now ends.
-> The notification is not received because subscribe() finished after
   the notification was published. The local cache is not invalidated
   and contains old data.

Error handling:
Subscribe now rejects the promise if the request fails. Previously, the
request was retried but may have failed again resulting in an
unsuccessful subscriptions -> notification for that topic were never
received.

408087